Поправил пагинацию на курсорную, добавил сваггер, починил исключения

+++ app/Handlers/Comment/DeleteCommentHandler.php
@@ -2,13 +2,17 @@
 
 namespace App\Handlers\Comment;
 
+use App\Exceptions\CommentNotFoundException;
 use App\Models\Comment;
 
 class DeleteCommentHandler
 {
     public function handle(int $id): void
     {
-        $comment = Comment::findOrFail($id);
+        $comment = Comment::find($id);
+        if (!$comment) {
+            throw new CommentNotFoundException($id);
+        }
         $comment->delete();
     }
 }
